requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
pillow>=9.0.0
numpy>=1.21.0